automatically generated arbitrary commit message for convenience

+++ ml/train.py
@@ -121,10 +121,10 @@ def model(x, num_variables, num_classes, fold, reuse=False):
 def model(x, num_variables, num_classes, fold, reuse=False):
     hidden_nodes = 100
     with tf.variable_scope('model_fold{}'.format(fold), reuse=reuse):
-        w1 = tf.get_variable('w1', shape=(num_variables, hidden_nodes), initializer=tf.random_normal_initializer())
-        b1 = tf.get_variable('b1', shape=(hidden_nodes), initializer=tf.constant_initializer())
-        w2 = tf.get_variable('w2', shape=(hidden_nodes, 1), initializer=tf.random_normal_initializer())
-        b2 = tf.get_variable('b2', shape=(1), initializer=tf.constant_initializer())
+        w1 = tf.get_variable('w1', shape=(num_variables, hidden_nodes), initializer=tf.random_normal_initializer(), dtype=tf.float64)
+        b1 = tf.get_variable('b1', shape=(hidden_nodes), initializer=tf.constant_initializer(), dtype=tf.float64)
+        w2 = tf.get_variable('w2', shape=(hidden_nodes, 1), initializer=tf.random_normal_initializer(), dtype=tf.float64)
+        b2 = tf.get_variable('b2', shape=(1), initializer=tf.constant_initializer(), dtype=tf.float64)
 
     l1 = tf.tanh(tf.add(b1, tf.matmul(x, w1)))
     logits = tf.add(b2, tf.matmul(l1, w2))
